First Year/1st Semester/Laboratory Of Advanced Programming/1986101_FORTUNA/source/backend-api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.31.0
psycopg2-binary==2.9.9
pydantic==2.9.2